refactoring: util 폴더 분리
